moves code that writes cc residual and eom ccsd python functions to file into pdaggerq/numerical_utils/autogen.py
